+++ Libft/ft_memcpy.c
@@ -6,13 +6,14 @@
 /*   By: ilorenzo <ilorenzo@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 15:00:22 by ilorenzo          #+#    #+#             */
-/*   Updated: 2023/09/15 15:51:15 by ilorenzo         ###   ########.fr       */
+/*   Updated: 2023/09/17 16:46:59 by ilorenzo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*cdest;
 	char	*csrc;
